fixing args for API and CLI
