+++ lib/budo.js
@@ -2,7 +2,6 @@ var Emitter = require('events/')
 var xtend = require('xtend')
 var assign = require('xtend/mutable')
 var log = require('bole')('budo')
-var dargs = require('dargs')
 
 var createServer = require('./server')
 var createFileWatch = require('./file-watch')
@@ -26,7 +25,7 @@ module.exports = function() {
   var defaultWatchOpts
   var deferreds = []
 
-  emitter._start = function(entries, opt) {
+  emitter._start = function(entries, opt, cli) {
     opt = opt || {}
     var port = opt.port
     server = createServer(opt)
@@ -52,7 +51,7 @@ module.exports = function() {
               : opt['ignore-watch']
         defaultWatchOpts = {
           poll: opt.poll,
-          'ignore-watch': ignoreWatchArg
+          'ignore-watch': opt.ignoreWatch || ignoreWatchArg
         }
         //and default live options
         defaultLiveOpts = {
@@ -62,7 +61,7 @@ module.exports = function() {
         }
 
         //start watchify process
-        runWatchify(entries, opt)
+        runWatchify(entries, opt, cli)
 
         serveAs = opt.serve
         started = true
@@ -144,15 +143,13 @@ module.exports = function() {
 
   return emitter
 
-  function runWatchify(entries, opt) {    
+  function runWatchify(entries, opt, cli) {    
     if (closed)
       return
 
-    //get watchify args (with --debug by default)
-    var watchifyArgs = getWatchifyArgs(entries, opt)
-
     //create a new watchify instance
-    watchify = createWatchify(watchifyArgs, { 
+    watchify = createWatchify(entries, opt, { 
+      cli: cli, //whether to use watchify/bin/args or not
       dir: opt.dir,
       serve: opt.serve,
       verbose: opt.v || opt.verbose,
@@ -193,34 +190,4 @@ function getChokidarOpts(opt) {
     delete opt.ignored
   }
   return opt
-}
-
-function getWatchifyArgs(entries, opt) {
-  //do not mutate original opts
-  opt = assign({}, opt)
-
-  //disable delay since we will handle debouncing manually
-  opt.delay = 0
-
-  //enable debug by default
-  if (opt.d !== false && opt.debug !== false) {
-    delete opt.d
-    opt.debug = true
-  }
-  //if user explicitly disabled debug...
-  else if (opt.d === false || opt.debug === false) {
-    delete opt.d
-    delete opt.debug
-  }
-
-  //clean up some possible collisions
-  var collisions = [
-    'dir', 'o', 'outfile', 'port',
-    'host', 'live', 'serve', 'live-port',
-    'live-plugin', 'defaultIndex'
-  ]
-  collisions.forEach(function(col) {
-    delete opt[col]
-  })
-  return entries.concat(dargs(opt))
 }
