+++ lib/watchify.js
@@ -1,11 +1,11 @@
 var log = require('bole')('budo')
-var getWatchify = require('./get-watchify')
+var createWatchify = require('./create-watchify')
 var Emitter = require('events/')
 var debounce = require('debounce')
 var concat = require('concat-stream')
 
 //Eventually this may split into a watchify-server module
-module.exports = function(watchifyArgs, opt) {
+module.exports = function(entries, userArgs, opt) {
   var emitter = new Emitter()
   var delay = opt.delay
   var closed = false
@@ -29,7 +29,10 @@ module.exports = function(watchifyArgs, opt) {
     }
   }
 
-  getWatchify({ basedir: opt.dir }, function(err, fromArgs) {
+  createWatchify(entries, userArgs, { 
+    cli: opt.cli,
+    basedir: opt.dir
+  }, function(err, instance) {
     if (err) { 
       var msg = [
         err.message,
@@ -42,8 +45,7 @@ module.exports = function(watchifyArgs, opt) {
     if (closed)
       return 
 
-    watchify = fromArgs(watchifyArgs)
-
+    watchify = instance
     var bundleDebounced = debounce(bundle, delay)
     watchify.on('update', function() {
       emitter.emit('pending')
